requirements.txt
pyowm==3.3.0
pyserial==3.5
